integration tests: move routing integration tests to a separate file
